requirements.txt
Flask
Flask-SQLAlchemy
scikit-learn
pandas
python-dotenv
bcrypt
